requirements.txt
python_telegram_bot[ext]==20.2
python-a2s==1.3.0
